Move demo test to jUnit engine Kotlin assertions tests
